+++ src/app/components/cluster-dns-check-table/messages/own-domain/dns-zone/unresolvable.js
@@ -1,0 +1,13 @@
+/**
+ * Dummy component file for Ember classic-component.
+ *
+ * @author Jakub Liput
+ * @copyright (C) 2024 ACK CYFRONET AGH
+ * @license This software is released under the MIT license cited in 'LICENSE.txt'.
+ */
+
+import Component from '@ember/component';
+
+export default Component.extend({
+  tagName: '',
+});