+++ src/app/templates/login.js
@@ -1,1 +1,0 @@
-export { default } from 'onepanel-gui/templates/onepanel-login';